# ZBP-polish-api-2_1_2-swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "version" : "2_1_2",
    "title" : "Polish API",
    "description" : "Interface specification for services provided by third parties based on access to payment accounts. Prepared by the Polish Bank Association and its affiliates"
  },
  "host" : "apiHost",
  "basePath" : "/",
  "schemes" : [ "https" ],
  "consumes" : [ "application/json" ],
  "produces" : [ "application/json" ],
...
  "parameters" : {
...
    "acceptEncodingParam" : {
      "name" : "Accept-Encoding",
      "in" : "header",
      "required" : true,
      "type" : "string",
      "enum" : [ "gzip", "deflate" ],
      "description" : "Gzip, deflate"
    },
    "acceptLanguageParam" : {
      "name" : "Accept-Language",
      "in" : "header",
      "required" : true,
      "type" : "string",
      "description" : "Prefered language of response"
    },
    "acceptCharsetParam" : {
      "name" : "Accept-Charset",
      "in" : "header",
      "required" : true,
      "type" : "string",
      "enum" : [ "utf-8" ],
      "description" : "UTF-8"
    },
    "xjwsSignatureParam" : {
      "name" : "X-JWS-SIGNATURE",
      "in" : "header",
      "required" : true,
      "type" : "string",
      "description" : "Detached JWS signature of the body of the payload"
    },
    "xRequestIdParam" : {
      "name" : "X-REQUEST-ID",
      "in" : "header",
      "required" : true,
      "type" : "string",
      "description" : "Identyfikator żądania w formacie UUID (Wariant 1, Wersja 1), zgodnym ze standardem RFC 4122, nadawany przez TPP. Wartość musi być zgodna z parametrem requestId przekazywanym w ciele każdego żądania. / Request identifier using UUID format (Variant 1, Version 1), described in RFC 4122 standard, set by TPP. Value of this header must be the same as for the requestId param passed inside request payload."
    }
  },
  "paths" : {
...
    "/v2_1_2.1/accounts/v2_1_2.1/deleteConsent" : {
      "x-swagger-router-controller" : "ais",
      "post" : {
        "summary" : "Removes consent",
        "description" : "Removes consent",
        "operationId" : "deleteConsent",
        "parameters" : [ {
          "$ref" : "#/parameters/acceptEncodingParam"
        }, {
          "$ref" : "#/parameters/acceptLanguageParam"
        }, {
          "$ref" : "#/parameters/acceptCharsetParam"
        }, {
          "$ref" : "#/parameters/xjwsSignatureParam"
        }, {
          "$ref" : "#/parameters/xRequestIdParam"
        }, {
          "name" : "deleteConsentRequest",
          "in" : "body",
          "description" : "Data for delete Consent Request",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/DeleteConsentRequest"
          }
        } ],
        "tags" : [ "AIS" ],
        "responses" : {
          "204" : {
            "description" : "Success",
            "headers" : {
              "Content-Encoding" : {
                "type" : "string",
                "description" : "Gzip, deflate"
              }
            }
          },
          "400" : {
            "description" : "Bad request",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "401" : {
            "description" : "Unauthorized",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "403" : {
            "description" : "Forbidden",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "405" : {
            "description" : "Method Not Allowed",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "406" : {
            "description" : "Not Acceptable",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "415" : {
            "description" : "Unsupported Media Type",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "422" : {
            "description" : "Unprocessable entity",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "500" : {
            "description" : "Internal Server Error",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "501" : {
            "description" : "Not Implemented",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "503" : {
            "description" : "Service Unavailable",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          }
        }
      }
    },
    "/v2_1_2.1/accounts/v2_1_2.1/getAccounts" : {
      "x-swagger-router-controller" : "ais",
      "post" : {
        "summary" : "Get information about all user's payment account",
        "description" : "User identification based on access token",
        "operationId" : "getAccounts",
        "parameters" : [ {
          "$ref" : "#/parameters/authorizationParam"
        }, {
          "$ref" : "#/parameters/acceptEncodingParam"
        }, {
          "$ref" : "#/parameters/acceptLanguageParam"
        }, {
          "$ref" : "#/parameters/acceptCharsetParam"
        }, {
          "$ref" : "#/parameters/xjwsSignatureParam"
        }, {
          "$ref" : "#/parameters/xRequestIdParam"
        }, {
          "name" : "getAccountsRequest",
          "in" : "body",
          "description" : "Data for Accounts Request",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/AccountsRequest"
          }
        } ],
        "tags" : [ "AIS" ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "schema" : {
              "$ref" : "#/definitions/AccountsResponse"
            },
            "headers" : {
              "Content-Encoding" : {
                "type" : "string",
                "description" : "Gzip, deflate"
              },
              "X-JWS-SIGNATURE" : {
                "type" : "string",
                "description" : "Detached JWS signature of the body of the response"
              }
            }
          },
          "400" : {
            "description" : "Bad request",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "401" : {
            "description" : "Unauthorized",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "403" : {
            "description" : "Forbidden",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "405" : {
            "description" : "Method Not Allowed",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "406" : {
            "description" : "Not Acceptable",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "415" : {
            "description" : "Unsupported Media Type",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "422" : {
            "description" : "Unprocessable entity",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "429" : {
            "description" : "Request limit for the requested service exceeded",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "500" : {
            "description" : "Internal Server Error",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "501" : {
            "description" : "Not Implemented",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "503" : {
            "description" : "Service Unavailable",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          }
        },
        "security" : [ {
          "xs2a_auth_aspsp" : [ "ais-accounts" ]
        }, {
          "xs2a_auth_decoupled" : [ "ais-accounts" ]
        } ]
      }
    },
    "/v2_1_2.1/accounts/v2_1_2.1/getAccount" : {
      "x-swagger-router-controller" : "ais",
      "post" : {
        "summary" : "Get detailed information about user payment account",
        "description" : "User identification based on access token",
        "operationId" : "getAccount",
        "parameters" : [ {
          "$ref" : "#/parameters/authorizationParam"
        }, {
          "$ref" : "#/parameters/acceptEncodingParam"
        }, {
          "$ref" : "#/parameters/acceptLanguageParam"
        }, {
          "$ref" : "#/parameters/acceptCharsetParam"
        }, {
          "$ref" : "#/parameters/xjwsSignatureParam"
        }, {
          "$ref" : "#/parameters/xRequestIdParam"
        }, {
          "name" : "getAccountRequest",
          "in" : "body",
          "description" : "Data for Account Request",
          "required" : true,
          "schema" : {
            "$ref" : "#/definitions/AccountInfoRequest"
          }
        } ],
        "tags" : [ "AIS" ],
        "responses" : {
          "200" : {
            "description" : "Success",
            "schema" : {
              "$ref" : "#/definitions/AccountResponse"
            },
            "headers" : {
              "Content-Encoding" : {
                "type" : "string",
                "description" : "Gzip, deflate"
              },
              "X-JWS-SIGNATURE" : {
                "type" : "string",
                "description" : "Detached JWS signature of the body of the response"
              }
            }
          },
          "400" : {
            "description" : "Bad request",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "401" : {
            "description" : "Unauthorized",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "403" : {
            "description" : "Forbidden",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "405" : {
            "description" : "Method Not Allowed",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "406" : {
            "description" : "Not Acceptable",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "415" : {
            "description" : "Unsupported Media Type",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "422" : {
            "description" : "Unprocessable entity",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "429" : {
            "description" : "Request limit for the requested service exceeded",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "500" : {
            "description" : "Internal Server Error",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "501" : {
            "description" : "Not Implemented",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          },
          "503" : {
            "description" : "Service Unavailable",
            "schema" : {
              "$ref" : "#/definitions/Error"
            }
          }
        },
        "security" : [ {
          "xs2a_auth_aspsp" : [ "ais" ]
        }, {
          "xs2a_auth_decoupled" : [ "ais" ]
        } ]
      }
    },
...
  },
  "securityDefinitions" : {
    "xs2a_auth_aspsp" : {
      "type" : "oauth2",
      "flow" : "accessCode",
      "authorizationUrl" : "https://apiHost/v2_1_2.1/auth/v2_1_2.1/authorize",
      "tokenUrl" : "https://apiHost/v2_1_2.1/auth/v2_1_2.1/token",
      "scopes" : {
        "ais-accounts" : "Permission to retrieve the list of accounts through Account Information Service",
        "ais" : "Permission to execute methods of Account Information Service in context of particular account",
        "pis" : "Permission to initiate payments and check their statuses through Payment Initiation Service"
      }
    },
    "xs2a_auth_decoupled" : {
      "type" : "oauth2",
      "flow" : "accessCode",
      "authorizationUrl" : "https://apiHost/auth/authorizeExt",
      "tokenUrl" : "https://apiHost/v2_1_2.1/auth/v2_1_2.1/token",
      "scopes" : {
        "ais-accounts" : "Permission to retrieve the list of accounts through Account Information Service",
        "ais" : "Permission to execute methods of Account Information Service in context of particular account",
        "pis" : "Permission to initiate payments and check their statuses through Payment Initiation Service"
      }
    }
  },
  "definitions" : {
...
    "DeleteConsentRequest" : {
      "description" : "Klasa zapytania o usunięcie zgody na dostęp do AIS",
      "type" : "object",
      "required" : [ "requestHeader", "consentId" ],
      "properties" : {
        "requestHeader" : {
          "$ref" : "#/definitions/RequestHeaderWithoutToken"
        },
        "consentId" : {
          "type" : "string",
          "description" : "Identyfikator zgody / Consent ID"
        }
      },
      "xml" : {
        "name" : "DeleteConsentRequest"
      }
    },
...
    "Error" : {
      "description" : "Klasa zawierająca informacje o zwracanym błędzie / Error class",
      "properties" : {
        "responseHeader" : {
          "$ref" : "#/definitions/ResponseHeader"
        },
        "code" : {
          "type" : "string",
          "description" : "Kod błędu / Error code"
        },
        "message" : {
          "type" : "string",
          "description" : "Opis błędu / Error message"
        }
      }
    },
...
  }
}